+++ python/coinbase-agentkit/coinbase_agentkit/action_providers/__init__.py
@@ -2,6 +2,7 @@
 
 from .action_decorator import create_action
 from .action_provider import Action, ActionProvider
+from .allora.allora_action_provider import AlloraActionProvider, allora_action_provider
 from .basename.basename_action_provider import (
     BasenameActionProvider,
     basename_action_provider,
@@ -10,8 +11,13 @@ from .cdp.cdp_api_action_provider import CdpApiActionProvider, cdp_api_action_pr
 from .cdp.cdp_wallet_action_provider import CdpWalletActionProvider, cdp_wallet_action_provider
 from .compound.compound_action_provider import CompoundActionProvider, compound_action_provider
 from .erc20.erc20_action_provider import ERC20ActionProvider, erc20_action_provider
+from .hyperboliclabs.hyperbolic_action_provider import (
+    HyperbolicActionProvider,
+    hyperbolic_action_provider,
+)
 from .morpho.morpho_action_provider import MorphoActionProvider, morpho_action_provider
 from .pyth.pyth_action_provider import PythActionProvider, pyth_action_provider
+from .ssh.ssh_action_provider import SshActionProvider, ssh_action_provider
 from .superfluid.superfluid_action_provider import (
     SuperfluidActionProvider,
     superfluid_action_provider,
@@ -25,6 +31,8 @@ __all__ = [
     "Action",
     "ActionProvider",
     "create_action",
+    "AlloraActionProvider",
+    "allora_action_provider",
     "BasenameActionProvider",
     "basename_action_provider",
     "CdpApiActionProvider",
@@ -35,10 +43,14 @@ __all__ = [
     "compound_action_provider",
     "ERC20ActionProvider",
     "erc20_action_provider",
+    "HyperbolicActionProvider",
+    "hyperbolic_action_provider",
     "MorphoActionProvider",
     "morpho_action_provider",
     "PythActionProvider",
     "pyth_action_provider",
+    "SshActionProvider",
+    "ssh_action_provider",
     "SuperfluidActionProvider",
     "superfluid_action_provider",
     "TwitterActionProvider",