+++ python/create-onchain-agent/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Coinbase Create Onchain Agent Changelog
 
+<!-- towncrier release notes start -->
+
+## [0.3.0] - 2025-03-14
+
+### Added
+
+- Added OpenAI Agents SDK support ([#557](https://github.com/coinbase/agentkit/pull/557))
+- Added --beginner flag
+
+
+## [0.2.0] - 2025-03-07
+
+### Fixed
+
+- Fixed indentation bug in chatbot.py template
+
+### Added
+
+- Added revised network selection ([#498](https://github.com/coinbase/agentkit/pull/498))
+- Added smart wallet support ([#503](https://github.com/coinbase/agentkit/pull/503))
+
+
 ## [0.1.4] - 2025-02-28
 
 ### Added